fix target folder manifest
