+++ packages/cli/src/handlers/dbt/compile.ts
@@ -54,7 +54,8 @@ const optionsToArgs = (options: DbtCompileOptions): string[] =>
     }, []);
 export const dbtCompile = async (options: DbtCompileOptions) => {
     try {
-        const args = optionsToArgs(options);
+        const args = optionsToArgs({ ...options });
+        // args.push(` --target-path ${options.projectDir}/target `)
         GlobalState.debug(`> Running: dbt compile ${args.join(' ')}`);
         const { stdout, stderr } = await execa('dbt', ['compile', ...args]);
         console.error(stdout);
