+++ packages/cli/src/handlers/dbt/getDbtVersion.ts
@@ -25,13 +25,3 @@ export const isSupportedDbtVersion = (version: string) => {
 
     return false;
 };
-
-export enum DbtManifestVersion {
-    V7 = 'v7',
-    V9 = 'v9',
-}
-export const getDbtManifest = async (): Promise<DbtManifestVersion> => {
-    const version = await getDbtVersion();
-    if (version.startsWith('1.5.')) return DbtManifestVersion.V9;
-    return DbtManifestVersion.V7;
-};
